style: fix circular dependencies
